requirements.txt
selenium==4.21.0
requests==2.31.0
pandas==2.2.2
python-dotenv==1.0.1